Add files via upload
Necessary files are menu.html and the folder 'Parallax-flipping-cards'. Code hasn't been documented, so bring your issues to me.

From the root folder, the CSS & JS folders are necessary, along with the 'steak.jpg', which serves as page background image.

'About us' page will be added shortly, since it is low priority and almost completely similar implementation to the menu page. Code snippets for it are already in place.

Lots of unnecessary files, such as fonts, scripts and images not used in the pages and carreid over by the template are present. Skimm them off or migrate the useful stuff to the latest build of the project.